Add DataSourceQueryResultSetConverter
