+++ shardingsphere-proxy/shardingsphere-proxy-common/src/test/java/org/apache/shardingsphere/proxy/config/util/DataSourceParameterConverterTest.java
@@ -17,49 +17,19 @@
 
 package org.apache.shardingsphere.proxy.config.util;
 
-import com.zaxxer.hikari.HikariDataSource;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceParameter;
 import org.apache.shardingsphere.proxy.config.yaml.YamlDataSourceParameter;
 import org.junit.Test;
 
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Properties;
 
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
 
 public final class DataSourceParameterConverterTest {
-    
-    @Test
-    public void assertGetDataSourceParameterMap() {
-        Map<String, Object> props = new LinkedHashMap<>();
-        props.put("maximumPoolSize", 10);
-        props.put("connectionTimeout", "30000");
-        props.put("readOnly", false);
-        Map<String, DataSourceConfiguration> dataSourceConfigurationMap = new HashMap<>(2, 1);
-        DataSourceConfiguration dataSourceConfiguration0 = new DataSourceConfiguration(HikariDataSource.class.getName());
-        dataSourceConfiguration0.getProps().put("url", "jdbc:mysql://localhost:3306/demo_ds_0");
-        dataSourceConfiguration0.getProps().putAll(props);
-        dataSourceConfigurationMap.put("ds_0", dataSourceConfiguration0);
-        DataSourceConfiguration dataSourceConfiguration1 = new DataSourceConfiguration(HikariDataSource.class.getName());
-        dataSourceConfiguration1.getProps().put("jdbcUrl", "jdbc:mysql://localhost:3306/demo_ds_1");
-        dataSourceConfiguration1.getProps().putAll(props);
-        dataSourceConfigurationMap.put("ds_1", dataSourceConfiguration1);
-        Map<String, DataSourceParameter> actual = DataSourceParameterConverter.getDataSourceParameterMap(dataSourceConfigurationMap);
-        assertThat(actual.size(), is(2));
-        assertThat(actual.get("ds_0").getUrl(), is("jdbc:mysql://localhost:3306/demo_ds_0"));
-        assertThat(actual.get("ds_0").getMaxPoolSize(), is(10));
-        assertThat(actual.get("ds_0").getConnectionTimeoutMilliseconds(), is(30000L));
-        assertFalse(actual.get("ds_1").isReadOnly());
-        assertThat(actual.get("ds_1").getUrl(), is("jdbc:mysql://localhost:3306/demo_ds_1"));
-        assertThat(actual.get("ds_1").getMaxPoolSize(), is(10));
-        assertThat(actual.get("ds_1").getConnectionTimeoutMilliseconds(), is(30000L));
-        assertFalse(actual.get("ds_1").isReadOnly());
-    }
     
     @Test
     public void assertGetDataSourceConfigurationMap() {
